chore: refactor reference to simapp
