+++ go.mod
@@ -5,6 +5,7 @@ go 1.20
 require (
 	cosmossdk.io/errors v1.0.0
 	cosmossdk.io/math v1.1.3-rc.1
+	cosmossdk.io/simapp v0.0.0-20231005184908-69d41f1bd51b
 	filippo.io/edwards25519 v1.0.0
 	github.com/SigmaGmbH/librustgo v0.0.2-0.20230725120829-f1a254f14ac6
 	github.com/armon/go-metrics v0.4.1
